+++ src/dcm/pkg/module/logicalcloud.go
@@ -48,11 +48,6 @@ type LogicalCloudKey struct {
 	LogicalCloudName string `json:"logicalCloud"`
 }
 
-// AppContextKey is an alternative key to access logical clouds
-type AppContextKey struct {
-	LCContext string `json:"logicalCloudContext"`
-}
-
 // LogicalCloudManager is an interface that exposes the connection
 // functionality
 type LogicalCloudManager interface {
@@ -67,20 +62,18 @@ type LogicalCloudManager interface {
 // LogicalCloudClient implements the LogicalCloudManager
 // It will also be used to maintain some localized state
 type LogicalCloudClient struct {
-	storeName  string
-	tagMeta    string
-	tagContext string
-	tagState   string
+	storeName string
+	tagMeta   string
+	tagState  string
 }
 
 // LogicalCloudClient returns an instance of the LogicalCloudClient
 // which implements the LogicalCloudManager
 func NewLogicalCloudClient() *LogicalCloudClient {
 	return &LogicalCloudClient{
-		storeName:  "resources",
-		tagMeta:    "data",
-		tagContext: "logicalCloudContext",
-		tagState:   "stateInfo",
+		storeName: "resources",
+		tagMeta:   "data",
+		tagState:  "stateInfo",
 	}
 }
 
@@ -109,7 +102,7 @@ func (v *LogicalCloudClient) Create(project string, c LogicalCloud) (LogicalClou
 		return LogicalCloud{}, pkgerrors.Wrap(err, "Creating DB Entry")
 	}
 
-	// Add the state info  record
+	// Add the state info record (initial state)
 	s := state.StateInfo{}
 	a := state.ActionEntry{
 		State:     state.StateEnum.Created,
@@ -226,9 +219,15 @@ func (v *LogicalCloudClient) Delete(project, logicalCloudName string) error {
 		return err
 	}
 
-	context, _, err := GetLogicalCloudContext(v.storeName, key, v.tagContext, project, logicalCloudName)
-	// If there's no context for Logical Cloud, just go ahead and delete it now
+	// Check if there was a previous context for this logical cloud
+	s, err := v.GetState(project, logicalCloudName)
 	if err != nil {
+		return err
+	}
+	cid := state.GetLastContextIdFromStateInfo(s)
+
+	// If there's no context for Logical Cloud, just go ahead and delete it now
+	if cid == "" {
 		err = db.DBconn.Remove(v.storeName, key)
 		if err != nil {
 			return pkgerrors.Wrap(err, "Error when deleting Logical Cloud (scenario with no context)")
@@ -236,9 +235,14 @@ func (v *LogicalCloudClient) Delete(project, logicalCloudName string) error {
 		return nil
 	}
 
+	ac, err := state.GetAppContextFromId(cid)
+	if err != nil {
+		return err
+	}
+
 	// Make sure rsync status for this logical cloud is Terminated,
-	// otherwise we can't remove appcontext yet
-	acStatus, err := GetAppContextStatus(context)
+	// otherwise we can't re-instantiate logical cloud yet
+	acStatus, err := GetAppContextStatus(ac)
 	if err != nil {
 		return err
 	}
@@ -261,7 +265,7 @@ func (v *LogicalCloudClient) Delete(project, logicalCloudName string) error {
 		fallthrough
 	case appcontext.AppContextStatusEnum.Terminated:
 		// remove the appcontext
-		err := context.DeleteCompositeApp()
+		err := ac.DeleteCompositeApp()
 		if err != nil {
 			log.Error("Error deleting AppContext CompositeApp Logical Cloud", log.Fields{"logicalcloud": logicalCloudName})
 			return pkgerrors.Wrap(err, "Error deleting AppContext CompositeApp Logical Cloud")
@@ -301,68 +305,6 @@ func (v *LogicalCloudClient) Update(project, logicalCloudName string, c LogicalC
 		return LogicalCloud{}, pkgerrors.Wrap(err, "Updating DB Entry")
 	}
 	return c, nil
-}
-
-// GetLogicalCloudContext returns the AppContext for corresponding provider and name
-func GetLogicalCloudContext(storeName string, key db.Key, meta string, project string, name string) (appcontext.AppContext, string, error) {
-
-	value, err := db.DBconn.Find(storeName, key, meta)
-	if err != nil {
-		return appcontext.AppContext{}, "", err
-	}
-
-	if len(value) == 0 {
-		return appcontext.AppContext{}, "", pkgerrors.New("Logical Cloud AppContext not found")
-	}
-
-	//value is a [][]byte
-	if value != nil {
-		ctxVal := string(value[0])
-		var lcc appcontext.AppContext
-		_, err = lcc.LoadAppContext(ctxVal)
-		if err != nil {
-			return appcontext.AppContext{}, "", pkgerrors.Wrap(err, "Reinitializing Logical Cloud AppContext")
-		}
-		return lcc, ctxVal, nil
-	}
-
-	return appcontext.AppContext{}, "", pkgerrors.New("Unknown Error")
-}
-
-// GetLogicalCloudFromContext returns the pair (project, logical cloud name) for a given AppContext
-func GetLogicalCloudFromContext(storeName string, appContextID string) (string, string, error) {
-	key := AppContextKey{
-		LCContext: appContextID,
-	}
-	log.Info("GetLogicalCloudFromContext", log.Fields{"appContextID": appContextID})
-
-	values, err := db.DBconn.Find(storeName, key, "logicalCloud")
-	if err != nil {
-		log.Error("Couldn't fetch logical cloud", log.Fields{"err": err})
-		return "", "", err
-	}
-
-	if len(values) == 0 {
-		return "", "", pkgerrors.New("Logical Cloud not found")
-	}
-
-	logicalCloudName := string(values[0])
-	log.Info("", log.Fields{"logicalCloudName": logicalCloudName})
-
-	values, err = db.DBconn.Find(storeName, key, "project")
-	if err != nil {
-		log.Error("Couldn't fetch project", log.Fields{"err": err})
-		return "", "", err
-	}
-
-	if len(values) == 0 {
-		return "", "", pkgerrors.New("Project not found")
-	}
-
-	project := string(values[0])
-	log.Info("", log.Fields{"project": project})
-
-	return project, logicalCloudName, nil
 }
 
 // GetAppContextStatus returns the Status for a particular AppContext
